continue transport stream work
